requirement.txt
absl-py==2.1.0
ale-py==0.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
box2d-py==2.3.5
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.2.1
cycler==0.12.1
debugpy @ file:///D:/bld/debugpy_1732236656051/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
etils==1.8.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.51.0
fsspec==2024.3.1
glfw==2.7.0
gymnasium==0.29.1
idna==3.7
imageio==2.34.0
importlib_metadata==8.4.0
importlib_resources==6.4.0
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1732896960890/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1731317204262/work
Jinja2==3.1.4
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///D:/bld/jupyter_core_1710257313664/work
kiwisolver==1.4.5
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mpmath==1.3.0
mujoco==3.1.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.4.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
psutil @ file:///D:/bld/psutil_1729847175714/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pycocotools==2.0.8
pygame==2.5.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1733221634316/work
PyOpenGL==3.1.7
pyparsing==3.1.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-docx==1.1.2
pytz==2024.1
pywin32==307
pyzmq @ file:///D:/bld/pyzmq_1728642364379/work
requests==2.31.0
scikit-learn==1.6.0
scipy==1.14.0
seaborn==0.13.2
Shimmy==0.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733216807016/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
swig==4.2.1
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
tornado @ file:///D:/bld/tornado_1732615900055/work
tqdm==4.66.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
zipp==3.18.1
